Make GameEnvironment into a Singleton class

See merge request seng201-2024/team-53!2

+++ src/main/java/seng201/team53/items/Purchasable.java
@@ -7,7 +7,7 @@ public interface Purchasable {
 
     int getCostPrice();
 
-    Double getSellPrice();
+    int getSellPrice();
 
     Boolean isSellable();
 }
