+++ src/main/java/seng201/team53/items/towers/LumberMillTower.java
@@ -25,8 +25,8 @@ public class LumberMillTower extends Tower {
     }
 
     @Override
-    public Double getSellPrice() {
-        return 50.0;
+    public int getSellPrice() {
+        return 50;
     }
 
     @Override
