+++ src/main/java/seng201/team53/items/towers/MineTower.java
@@ -24,8 +24,8 @@ public class MineTower extends Tower {
     }
 
     @Override
-    public Double getSellPrice() {
-        return 50.0;
+    public int getSellPrice() {
+        return 50;
     }
 
     @Override
